backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "edgartools>=5.13.0",
    "fastapi>=0.128.0",
    "google-genai>=1.61.0",
    "html-to-markdown>=2.24.1",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "langchain-text-splitters>=1.1.0",
    "lxml>=6.0.2",
    "markitdown>=0.1.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sec-parser>=0.1.0",
    "sqlalchemy>=2.0.46",
    "supabase>=2.27.2",
    "uvicorn[standard]>=0.40.0",
    "weaviate-client>=4.19.2",
]
